# Optimize deriveshaOrig