--- a/tests/tx_root_test.go
+++ b/tests/tx_root_test.go
@@ -18,13 +18,13 @@ package tests
 
 import (
 	"fmt"
-	"github.com/klaytn/klaytn/crypto"
 	"math/big"
 	"testing"
 	"time"
 
 	"github.com/klaytn/klaytn/blockchain/types"
 	"github.com/klaytn/klaytn/common/profile"
+	"github.com/klaytn/klaytn/crypto"
 	"github.com/klaytn/klaytn/storage/statedb"
 )
 
@@ -116,7 +116,7 @@ func genTxs(num uint64) (types.Transactions, error) {
 	if err != nil {
 		return nil, err
 	}
-	var addr = crypto.PubkeyToAddress(key.PublicKey)
+	addr := crypto.PubkeyToAddress(key.PublicKey)
 	newTx := func(i uint64) (*types.Transaction, error) {
 		signer := types.NewEIP155Signer(big.NewInt(18))
 		utx := types.NewTransaction(i, addr, new(big.Int), 0, new(big.Int).SetUint64(10000000), nil)