@kjhman21 2022-10-25
@hqjang-pepper Can you share the performance result of the benchmark here as a comment?