@hqjang-pepper 2022-10-25
Tested with 100, 1000, 4000 txs, using StackTrie

```
Before this PR :
BenchmarkDeriveShaOrig/1000test-stackTrie-10         	    7971	    153447 ns/op
BenchmarkDeriveShaOrig/1000test-stackTrie-10         	     793	   1499944 ns/op
BenchmarkDeriveShaOrig/4000test-stackTrie-10         	     195	   5928197 ns/op

After this PR :
BenchmarkDeriveShaOrig/100test-stackTrie-10         	    7995	    149334 ns/op. -  2.7% faster
BenchmarkDeriveShaOrig/1000test-stackTrie-10         	     788	   1488989 ns/op. -  0.7% faster
BenchmarkDeriveShaOrig/4000test-stackTrie-10         	     199	   5891009 ns/op. -  0.6% faster

```
2.7%