@hqjang-pepper 2022-10-25
Tested with 100, 1000, 4000 txs, using StackTrie

Before this PR : 
BenchmarkDeriveShaOrig/100test-stackTrie-10         	    7977	    148040 ns/op
BenchmarkDeriveShaOrig/1000test-stackTrie-10         	     744	   1533285 ns/op
BenchmarkDeriveShaOrig/1000test-stackTrie-10         	     194	   5926051 ns/op


After this PR :
BenchmarkDeriveShaOrig/100test-stackTrie-10         	    8449	    140057 ns/op  -  5.7% faster
BenchmarkDeriveShaOrig/1000test-stackTrie-10         	     836	   1441072 ns/op. -  6.4% faster
BenchmarkDeriveShaOrig/4000test-stackTrie-10         	     210	   5555414 ns/op. -  6.6% faster